+++ app/src/androidTest/java/se/sugarest/jane/viaplaysections/mainActivityTests/HasContentMainActivityTest.java
@@ -13,7 +13,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import se.sugarest.jane.viaplaysections.R;
-import se.sugarest.jane.viaplaysections.ui.MainActivity;
+import se.sugarest.jane.viaplaysections.ui.list.MainActivity;
 import se.sugarest.jane.viaplaysections.util.DrawableMatcher;
 
 import static android.support.test.espresso.Espresso.onView;
