+++ app/src/main/res/layout/activity_main.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="se.sugarest.jane.viaplaysections.ui.MainActivity">
+    tools:context="se.sugarest.jane.viaplaysections.ui.list.MainActivity">
 
     <android.support.v4.widget.DrawerLayout
         android:id="@+id/drawer_layout"
