+++ app/src/main/res/layout/fragment_section_detail.xml
