+++ app/src/test/java/se/sugarest/jane/viaplaysections/SectionNamesFilterTest.java
@@ -10,7 +10,7 @@
 //import java.util.ArrayList;
 //import java.util.stream.Stream;
 //
-//import se.sugarest.jane.viaplaysections.ui.MainActivity;
+//import se.sugarest.jane.viaplaysections.ui.list.MainActivity;
 //
 ///**
 // * Created by jane on 17-11-27.
